Refactor starter template to use new RSC convention
