+++ templates/template-hydrogen-default/src/components/CartIconWithItems.jsx
