+++ templates/template-hydrogen-default/src/components/CartProvider.jsx
@@ -1,7 +1,9 @@
+'client export';
+
 import {useCallback} from 'react';
 import {CartProvider as ShopifyCartProvider} from '@shopify/hydrogen/client';
 
-import CartUIProvider, {useCartUI} from './CartUIProvider.client';
+import CartUIProvider, {useCartUI} from './CartUIProvider';
 
 /**
  * A client component that creates a cart object and provides callbacks that can be accessed by any descendent component using the `useCart` hook and related hooks
