+++ templates/template-hydrogen-default/src/components/Footer.jsx
