+++ templates/template-hydrogen-default/src/components/Gallery.jsx
