+++ templates/template-hydrogen-default/src/components/Header.jsx
@@ -1,11 +1,13 @@
+'client export';
+
 import {useEffect, useState} from 'react';
 import {Link} from '@shopify/hydrogen/client';
 
-import CartToggle from './CartToggle.client';
-import {useCartUI} from './CartUIProvider.client';
-import CountrySelector from './CountrySelector.client';
-import Navigation from './Navigation.client';
-import MobileNavigation from './MobileNavigation.client';
+import CartToggle from './CartToggle';
+import {useCartUI} from './CartUIProvider';
+import CountrySelector from './CountrySelector';
+import Navigation from './Navigation';
+import MobileNavigation from './MobileNavigation';
 
 /**
  * A client component that specifies the content of the header on the website
