+++ templates/template-hydrogen-default/src/components/Layout.jsx
@@ -7,9 +7,9 @@ import {
 } from '@shopify/hydrogen';
 import gql from 'graphql-tag';
 
-import Header from './Header.client';
-import Footer from './Footer.server';
-import Cart from './Cart.client';
+import Header from './Header';
+import Footer from './Footer';
+import Cart from './Cart';
 import {Suspense} from 'react';
 
 /**
