+++ templates/template-hydrogen-default/src/components/LoadMoreProducts.jsx
@@ -1,5 +1,7 @@
+'client export';
+
 import {useServerProps} from '@shopify/hydrogen/client';
-import SpinnerIcon from './SpinnerIcon.client';
+import SpinnerIcon from './SpinnerIcon';
 
 /**
  * A client component that provides functionality to initially show a subset of products and a button to load more products
