+++ templates/template-hydrogen-default/src/components/MoneyPrice.jsx
