+++ templates/template-hydrogen-default/src/components/SpinnerIcon.jsx
