+++ templates/template-hydrogen-default/src/components/Welcome.jsx
