+++ templates/template-hydrogen-default/src/routes/collections/[handle].jsx
@@ -1,10 +1,10 @@
 import {useShop, useShopQuery, flattenConnection, Seo} from '@shopify/hydrogen';
 import gql from 'graphql-tag';
 
-import LoadMoreProducts from '../../components/LoadMoreProducts.client';
-import Layout from '../../components/Layout.server';
+import LoadMoreProducts from '../../components/LoadMoreProducts';
+import Layout from '../../components/Layout';
 import ProductCard from '../../components/ProductCard';
-import NotFound from '../../components/NotFound.server';
+import NotFound from '../../components/NotFound';
 
 export default function Collection({
   country = {isoCode: 'US'},
