+++ templates/template-hydrogen-default/src/routes/countries.jsx
