+++ templates/template-hydrogen-default/src/routes/redirect.jsx
