+++ templates/template-hydrogen-default/src/routes/robots.txt.js
