+++ templates/template-hydrogen-default/src/routes/sitemap.xml.jsx
